refactor(mps-sync-plugin-lib): move INodeExt to a more appropriate package
